requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.18.2
paho-mqtt==1.5.0
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
